// src/types.ts
export interface Diagnosis {
    code: string;
    name: string;
    latin?: string;
  }

  export interface BaseEntry {
    id: string;
    description: string;
    date: string;
    specialist: string;
    diagnosisCodes?: Array<Diagnosis['code']>;
  }

  export enum HealthCheckRating {
    "Healthy" = 0,
    "LowRisk" = 1,
    "HighRisk" = 2,
    "CriticalRisk" = 3
  }

  interface SickLeave {
    startDate: string;
    endDate: string;
  }

  interface Discharge {
      date: string;
      criteria: string;
  }
  
  export interface HealthCheckEntry extends BaseEntry {
    type: "HealthCheck";
    healthCheckRating: HealthCheckRating;
  }

  export interface HospitalEntry extends BaseEntry {
    type: "Hospital";
    discharge: Discharge;
  }

  export interface OccupationalHealthcareEntry extends BaseEntry {
    type: "OccupationalHealthcare";
    employerName: string;
    sickLeave?: SickLeave;
  }
  
  export type Entry =
  | HospitalEntry
  | OccupationalHealthcareEntry
  | HealthCheckEntry;


export interface Patient {
  id: string,
  name: string,
  dateOfBirth: string,
  ssn: string,
  gender: Gender,
  occupation: string,
  entries: Entry[]
}



export type PublicPatient = Omit<Patient, 'ssn' | 'entries' >

export type NewPatientEntry = Omit<Patient, 'id' >;

export type NewEntry = Omit<Entry, 'id'>;

export enum Gender {
  Male = 'male',
  Female = 'female',
  Other = 'other' 
}